+++ Leomard/AppStorage/UserPreferences.swift
@@ -29,22 +29,25 @@ final class UserPreferences: ObservableObject {
     @AppStorage("unreadonly_when_opening_inbox", store: .standard) var unreadonlyWhenOpeningInbox: Bool = true
     @AppStorage("profile_sort_method", store: .standard) var profileSortMethod: SortType = .new
     @AppStorage("check_for_update_frequency", store: .standard) var checkForUpdateFrequency: UpdateFrequency = .onceADay
+    
     // MARK: - Mark as Read Stuff
     @AppStorage("mark_post_as_read_on_open", store: .standard) var markPostAsReadOnOpen: Bool = true
     @AppStorage("mark_post_as_read_on_vote", store: .standard) var markPostAsReadOnVote: Bool = true
     @AppStorage("mark_post_as_read_on_disappear", store: .standard) var markPostAsReadOnDisappear: Bool = false
     @AppStorage("hide_read_posts", store: .standard) var hideReadPosts: Bool = false
     
-    // MARK: - Not in Preferences.
-    @AppStorage("skipped_update_version", store: .standard) var skippedUpdateVersion: String = ""
-    @AppStorage("last_update_check_date", store: .standard) var lastUpdateCheckDate: Date = Date()
-    
     // MARK: - Blocked instances
     @AppStorage("blocked_instances", store: .standard) var blockedInstances: String = ""
     
-    // MARK: - Post View
+    // MARK: - Post
     @AppStorage("use_post_compact_view", store: .standard) var usePostCompactView: Bool = false
-    @AppStorage("two_column_view", store: .standard) var twoColumnView: Bool = false
+    @AppStorage("two_column_view", store: .standard) var twoColumnView: Bool = true
+    @AppStorage("save_liked_posts", store: .standard) var saveLikedPosts: Bool = false
+    @AppStorage("autosave_post_creation", store: .standard) var autosavePostCreation: Bool = true
+    @AppStorage("use_piped_video_for_youtube", store: .standard) var usePipedVideoForYoutube: Bool = false
+    @AppStorage("truncate_post_titles", store: .standard) var truncatePostTitles: Bool = true
+    @AppStorage("grayout_read_posts", store: .standard) var grayoutReadPosts: Bool = true
+    @AppStorage("show_read_post_indicator", store: .standard) var showReadPostIndicator: Bool = false
     
     // MARK: - Navbar
     @AppStorage("navbar_show_letter_separators", store: .standard) var navbarShowLetterSeparators: Bool = false
@@ -54,6 +57,13 @@ final class UserPreferences: ObservableObject {
     @AppStorage("prefer_display_name_community_post", store: .standard) var preferDisplayNameCommunityPost: Bool = false
     @AppStorage("prefer_display_name_communtiy_followed", store: .standard) var preferDisplayNameCommunityFollowed: Bool = false
     @AppStorage("prefer_display_name_people_post", store: .standard) var preferDisplayNamePeoplePost: Bool = false
+        
+    
+    
+    
+    // MARK: - Not in Preferences.
+    @AppStorage("skipped_update_version", store: .standard) var skippedUpdateVersion: String = ""
+    @AppStorage("last_update_check_date", store: .standard) var lastUpdateCheckDate: Date = Date()
     
     let sortTypes: [SortType] = [ .topHour, .topDay, .topMonth, .topYear, .hot, .active, .new, .mostComments ]
     let profileSortTypes: [SortType] = [ .topWeek, .topMonth, .topYear, .hot, .active, .new, .mostComments, .old ]